more refactoring

+++ game/Datatypes/TesObjectCell.h
@@ -11,7 +11,7 @@ public:
 	char padding0008[0x18];
 	std::uint32_t formId;
 	char padding0024[0x40];
-	std::uintptr_t isInterior;
+	std::uint64_t isInterior;
 	char loadedState;
 	char padding0071[0x2F];
 	std::uintptr_t objectListBeginPtr;//0xA0
